+++ wsidicom/file/io/frame_index/native_pixel_data.py
@@ -17,13 +17,13 @@
 import math
 from typing import List, Tuple
 
-from wsidicom.file.io.frame_index.frame_index import FrameIndex
 from wsidicom.file.io.frame_index.offset_table_type import OffsetTableType
+from wsidicom.file.io.frame_index.parser import FrameIndexParser
 from wsidicom.file.io.wsidicom_io import WsiDicomIO
 from wsidicom.geometry import Size
 
 
-class NativePixelData(FrameIndex):
+class NativePixelDataFrameIndexParser(FrameIndexParser):
     def __init__(
         self,
         file: WsiDicomIO,